# TheBoard.API/swagger.json
{
  "x-generator": "NSwag v14.6.3.0 (NJsonSchema v11.5.2.0 (Newtonsoft.Json v13.0.0.0))",
  "openapi": "3.0.0",
  "info": {
    "title": "TheBoard API",
    "version": "v1"
  },
  "paths": {
    "/v1/BoardItem/BoardItems/{projectId}": {
      "get": {
        "tags": [
          "BoardItem"
        ],
        "operationId": "BoardItem_GetBoardItemsByProjectId",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BoardItemVM"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/BoardItem": {
      "post": {
        "tags": [
          "BoardItem"
        ],
        "operationId": "BoardItem_CreateBoardItem",
        "requestBody": {
          "x-name": "boardItemCreateVM",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BoardItemCreateVM"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardItemVM"
                }
              }
            }
          }
        }
      }
    },
    "/v1/BoardItem/BatchUpdatePriority": {
      "post": {
        "tags": [
          "BoardItem"
        ],
        "operationId": "BoardItem_UpdateBoardItemPriority",
        "requestBody": {
          "x-name": "updates",
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/BoardItemPriorityUpdateVM"
                }
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/v1/BoardItem/MarkAsCompleted/{boardItemId}": {
      "post": {
        "tags": [
          "BoardItem"
        ],
        "operationId": "BoardItem_MarkBoardItemAsCompleted",
        "parameters": [
          {
            "name": "boardItemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/v1/BoardItem/{boardItemId}": {
      "delete": {
        "tags": [
          "BoardItem"
        ],
        "operationId": "BoardItem_DeleteBoardItem",
        "parameters": [
          {
            "name": "boardItemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/v1/Project/Projects": {
      "get": {
        "tags": [
          "Project"
        ],
        "operationId": "Project_GetAllProjects",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProjectVM"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/Project/{ProjectId}": {
      "get": {
        "tags": [
          "Project"
        ],
        "operationId": "Project_GetProjectById",
        "parameters": [
          {
            "name": "ProjectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectVM"
                }
              }
            }
          }
        }
      }
    },
    "/v1/Project": {
      "post": {
        "tags": [
          "Project"
        ],
        "operationId": "Project_CreateProject",
        "requestBody": {
          "x-name": "name",
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectVM"
                }
              }
            }
          }
        }
      }
    },
    "/v1/Project/{projectId}": {
      "delete": {
        "tags": [
          "Project"
        ],
        "operationId": "Project_DeleteProject",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/WeatherForecast": {
      "get": {
        "tags": [
          "WeatherForecast"
        ],
        "operationId": "WeatherForecast_Get",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WeatherForecast"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "BoardItemVM": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "title",
          "description",
          "createdAt",
          "projectId",
          "priority",
          "isCompleted"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "projectId": {
            "type": "integer",
            "format": "int32"
          },
          "priority": {
            "type": "integer",
            "format": "int32"
          },
          "isCompleted": {
            "type": "boolean"
          }
        }
      },
      "BoardItemCreateVM": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "title",
          "description",
          "projectId",
          "priority"
        ],
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "projectId": {
            "type": "integer",
            "format": "int32"
          },
          "priority": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "BoardItemPriorityUpdateVM": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "boardItemId",
          "newPriority"
        ],
        "properties": {
          "boardItemId": {
            "type": "integer",
            "format": "int32"
          },
          "newPriority": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "ProjectVM": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "projectId",
          "name"
        ],
        "properties": {
          "projectId": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "WeatherForecast": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "date",
          "temperatureC",
          "temperatureF"
        ],
        "properties": {
          "date": {
            "type": "string",
            "format": "date"
          },
          "temperatureC": {
            "type": "integer",
            "format": "int32"
          },
          "temperatureF": {
            "type": "integer",
            "format": "int32"
          },
          "summary": {
            "type": "string",
            "nullable": true
          }
        }
      }
    }
  }
}